+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [1.1.4] - 2021-12-16
+- added coft token fields in cards list api
+- fixed order refund api
+
 ## [1.1.3] - 2019-10-05
 ### Added
 - client_auth_token to Order